+++ libopenmpt/resource.h
@@ -12,6 +12,7 @@
 #define IDC_SLIDER_STEREOSEPARATION     1006
 #define IDC_COMBO_RAMPING               1007
 #define IDC_FILEINFO                    1008
+#define IDC_CHECK_AMIGA_RESAMPLER       1009
 
 // Next default values for new objects
 // 
@@ -19,7 +20,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        103
 #define _APS_NEXT_COMMAND_VALUE         40001
-#define _APS_NEXT_CONTROL_VALUE         1009
+#define _APS_NEXT_CONTROL_VALUE         1010
 #define _APS_NEXT_SYMED_VALUE           101
 #endif
 #endif